Split ReelHolder to ReelHolder and ReelHolderArm files and updated stl's to match.
Changed screw holes to support holding M3 nuts also so that the HeatFit inserts are not required for the reel holder top.

Changed screw hole (well nut holder) to be deeper than the others as the screw has to go through the top component block as well and it's a long distance for a M3 screw.